Refactor

+++ WorldWar.Abstractions/Extensions/BearingCalculator.cs
@@ -9,7 +9,7 @@ public static class BearingCalculator
 		var dLon = ToRad(lon2 - lon1);
 		var dPhi = Math.Log(
 			Math.Tan(ToRad(lat2) / 2 + Math.PI / 4) / Math.Tan(ToRad(lat1) / 2 + Math.PI / 4));
-		if (Math.Abs((double)dLon) > Math.PI)
+		if (Math.Abs(dLon) > Math.PI)
 			dLon = dLon > 0 ? -(2 * Math.PI - dLon) : (2 * Math.PI + dLon);
 		return ToBearing(Math.Atan2(dLon, dPhi));
 	}
